+++ packages/pangea-sdk/PangeaCyber.Net/Sanitize/Models/SanitizeFile.cs
@@ -1,0 +1,15 @@
+using Newtonsoft.Json;
+
+namespace PangeaCyber.Net.Sanitize
+{
+    ///
+    public class SanitizeFile
+    {
+        ///
+        [JsonProperty("scan_provider")]
+        public string? ScanProvider { get; set; }
+
+        ///
+        public SanitizeFile() { }
+    }
+}